behavioral-service/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0

# Machine learning
scikit-learn>=1.4.0
joblib>=1.3.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
hypothesis>=6.92.0
httpx>=0.26.0

# Logging
structlog>=24.1.0
